fixes to dataloading for campatibility with vae

+++ project/agou2_dusad2/code/dataset.py
@@ -6,11 +6,10 @@ from imageio import imread
 import os
 from constants import DATASET_SIZE, HEIGHT, WIDTH, ROLLOUT_DIR
 
-#%%
 class VAEDataset(Dataset):
     def __init__(self, rollout_dir=ROLLOUT_DIR, size=DATASET_SIZE, transform=None):
         self.transform=transform
-        self.frames = np.zeros((size, HEIGHT, WIDTH, 3), dtype=int)
+        self.frames = np.zeros((size, HEIGHT, WIDTH, 3), dtype=np.uint8)
         episodes = [ep for ep in os.listdir(ROLLOUT_DIR) if '.npz' in ep]
         num_frames = 0
         for episode in episodes:
